Separate file for print types
